loads of optimizations

+++ main/accelerometer/measurement.h
@@ -6,7 +6,7 @@ struct AccData
     int16_t x;
     int16_t y;
     int16_t z;
-    int64_t time;
+    unsigned int time;
 };
 
 struct AccDataDecimal
